// src/model/PlanDayDetail.ts
export class PlanDayDetail {
    TOUR_ID: number;
    SORT_DAY: number;
    LOCATION_DAY: number;
    LOCATION_ID: number;
    LOCATION_TYPE: number;
    STATUS: number;
    LOCATION_START_TIME: string;
    LOCATION_DESCRIPTION: string;
    LOCATION_GOOGLE: number;
    LOCATION_NAME: string;
    LOCATION_AVG_PRICE: string;
    LOCATION_AVATAR: string;
    LOCATION_CLOSE_TIME1: string;
    LOCATION_OPEN_TIME1: string;
    CITY_ID: number;
    NATION_ID: number;
    PROVINCE_ID: number;
}